+++ CHANGELOG.md
@@ -9,6 +9,14 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
+## [1.2.2] - 2023-11-28
+
+### Changed
+
+- Fixed issue where resource display was not set when rebuilding user booking entry cache.
+- Added "responseRequested" to booking invitations to fix issue with status.
+- Fixed status when resource has not answered event invitation in accept flow.
+
 ## [1.2.1] - 2023-11-23
 
 ### Changed